+++ CHANGELOG.md
@@ -15,3 +15,9 @@ NEXT:
 DONE:
 - Project Control System Started
 
+
+## 2026-07-15 01:19
+
+DONE:
+- Project Brain System Created
+